# doc/apidoc/schema_swagger_form_data.json
{
  "swagger": "2.0",
  "info": {
    "title": "CheckoutBasket (params in:formData)",
    "description": "\n<p>Endpoint docs</p>\n",
    "version": "1.0",
    "x-copyright": null
  },
  "basePath": "/",
  "consumes": [
    "application/x-www-form-urlencoded",
    "multipart/form-data"
  ],
  "paths": {
...
    "/admin/checkout_products/{id}": {
      "get": {
        "tags": [
          "checkout_products"
        ],
        "consumes": [
          "application/x-www-form-urlencoded",
          "multipart/form-data"
        ],
        "operationId": "get_admin_checkout_products_id",
        "summary": "Show a checkout product",
        "parameters": [
          {
            "name": "id",
            "type": "number",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "ok"
          }
        },
        "description": ""
      },
      "patch": {
        "tags": [
          "checkout_products"
        ],
        "consumes": [
          "application/x-www-form-urlencoded",
          "multipart/form-data"
        ],
        "operationId": "patch_admin_checkout_products_id",
        "summary": "Update a checkout product",
        "parameters": [
          {
            "name": "id",
            "type": "number",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "422": {
            "description": null
          }
        },
        "description": ""
      },
      "delete": {
        "tags": [
          "checkout_products"
        ],
        "consumes": [
          "application/x-www-form-urlencoded",
          "multipart/form-data"
        ],
        "operationId": "delete_admin_checkout_products_id",
        "summary": "Destroy a checkout product",
        "parameters": [
          {
            "name": "id",
            "type": "number",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "ok"
          }
        },
        "description": ""
      }
    },
...
    "/admin/products/{id}": {
      "get": {
        "tags": [
          "products"
        ],
        "consumes": [
          "application/x-www-form-urlencoded",
          "multipart/form-data"
        ],
        "operationId": "get_admin_products_id",
        "summary": "Show a product",
        "parameters": [
          {
            "name": "id",
            "type": "number",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "ok"
          }
        },
        "description": ""
      },
      "patch": {
        "tags": [
          "products"
        ],
        "consumes": [
          "application/x-www-form-urlencoded",
          "multipart/form-data"
        ],
        "operationId": "patch_admin_products_id",
        "summary": "Update a product",
        "parameters": [
          {
            "name": "id",
            "type": "number",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "422": {
            "description": null
          }
        },
        "description": ""
      },
      "delete": {
        "tags": [
          "products"
        ],
        "consumes": [
          "application/x-www-form-urlencoded",
          "multipart/form-data"
        ],
        "operationId": "delete_admin_products_id",
        "summary": "Destroy a product",
        "parameters": [
          {
            "name": "id",
            "type": "number",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "ok"
          }
        },
        "description": ""
      }
    },
...
  },
...
  "host": "localhost:3000"
}